split layers into files
