+++ espnet/nets/pytorch_backend/transformer/encoder_layer.py
@@ -1,0 +1,25 @@
+from torch import nn
+
+from .layer_norm import LayerNorm
+
+
+class EncoderLayer(nn.Module):
+    def __init__(self, size, self_attn, feed_forward, dropout):
+        super(EncoderLayer, self).__init__()
+        self.self_attn = self_attn
+        self.feed_forward = feed_forward
+        self.norm1 = LayerNorm(size)
+        self.norm2 = LayerNorm(size)
+        self.dropout = nn.Dropout(dropout)
+        self.size = size
+
+    def forward(self, x, mask):
+        """Compute encoded features
+
+        :param torch.Tensor x: encoded source features (batch, max_time_in, size)
+        :param torch.Tensor mask: mask for x (batch, max_time_in)
+        """
+        nx = self.norm1(x)
+        x = x + self.dropout(self.self_attn(nx, nx, nx, mask))
+        nx = self.norm2(x)
+        return x + self.dropout(self.feed_forward(nx)), mask
