+++ espnet/nets/pytorch_backend/transformer/layer_norm.py
@@ -1,0 +1,12 @@
+import torch
+
+
+class LayerNorm(torch.nn.LayerNorm):
+    def __init__(self, nout, dim=-1):
+        super(LayerNorm, self).__init__(nout, eps=1e-12)
+        self.dim = dim
+
+    def forward(self, x):
+        if self.dim == -1:
+            return super(LayerNorm, self).forward(x)
+        return super(LayerNorm, self).forward(x.transpose(1, -1)).transpose(1, -1)
